+++ CHANGELOG.md
@@ -7,10 +7,12 @@ adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+## [0.2.0] - 2020-08-30
+
 ### Added
 
 - A `Snek.Board.Snake.step/2` function for finding points relative to a snake's
-- last moved direction, but without moving the snake.
+  last moved direction, but without moving the snake.
 
 ## [0.1.0] - 2020-08-30
 
@@ -21,7 +23,8 @@ adheres to [Semantic Versioning][semver].
 - A `Snek.Ruleset.Standard` module implementing the Battlesnake Standard rules.
 - A `Snek.Ruleset.Solo` module implementing the Battlesnake Solo rules.
 
-[Unreleased]: https://github.com/xtagon/snek/compare/v0.1.0...edge
+[Unreleased]: https://github.com/xtagon/snek/compare/v0.2.0...edge
+[0.2.0]: https://github.com/xtagon/snek/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/xtagon/snek/releases/tag/v0.1.0
 
 [keepachangelog]: https://keepachangelog.com/en/1.0.0/